requirements.txt
asgiref==3.6.0
bidict==0.22.
gunicorn==20.1.0
h11==0.14.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-engineio==4.7.1
python-socketio==5.8.0
simple-websocket==0.10.1
wsproto==1.2.0

